+++ dsmr_backend/management/commands/dsmr_superuser.py
@@ -1,23 +1,18 @@
-import os
-import logging
-
 from django.contrib.auth.models import User
 from django.core.management.base import BaseCommand, CommandError
+from decouple import config
 
 
-class Command(BaseCommand):
+class Command(BaseCommand):  # noqa
     help = 'Persists or updates a superuser, as defined by the DSMR_USER and DSMR_PASSWORD environment vars.'
 
     def handle(self, **options):
         """ WARNING: Only safe for command line execution. Do NOT use for web requests! """
-        username = os.environ.get('DSMR_USER')
-        password = os.environ.get('DSMR_PASSWORD')
+        username = config('DSMR_USER')
+        password = config('DSMR_PASSWORD')
 
-        if not username:
-            raise CommandError('"DSMR_USER" not specified as environment variable')
-
-        if not password:
-            raise CommandError('"DSMR_PASSWORD" not specified as environment variable')
+        if not username or not password:
+            raise CommandError('DSMR_USER or DSMR_PASSWORD (or both) are empty')
 
         try:
             user = User.objects.get(
@@ -25,9 +20,9 @@ class Command(BaseCommand):
                 is_superuser=True
             )
         except User.DoesNotExist:
-            logging.info('Creating new superuser %s', username)
+            print('Creating new superuser "{}" (DSMR_USER)'.format(username))
             User.objects.create_superuser(username, '{}@localhost'.format(username), password)
         else:
-            logging.info('Updating password of superuser %s', username)
+            print('Updating password of superuser "{}" (DSMR_USER)'.format(username))
             user.set_password(password)
             user.save()