+++ dsmrreader/config/circleci/sqlite.py
@@ -1,7 +1,0 @@
-from dsmrreader.config.test.base import *
-
-DATABASES = {
-    'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-    }
-}