+++ tools/quick-test.sh
@@ -7,10 +7,11 @@ do
     ARGS="$ARGS $CURRENT"
 done
 
-
 echo ""
 echo "--- Testing with SQLite (4 processes)..."
-time pytest --cov --cov-report=html --cov-report=term --ds=dsmrreader.config.test.sqlite -n 2 $ARGS
+DB_ENGINE=django.db.backends.sqlite3
+export DB_ENGINE
+time pytest --cov --cov-report=html --cov-report=term --ds=dsmrreader.config.test -n 2 $ARGS
 
 
 echo ""